pyproject.toml
[tool.poetry]
name = "elcom"
version = "0.1.0"
description = ""
authors = ["Tamir Mandreev <tamirmandreev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
